requirements.txt
langchain
langchain-groq
langchain-openai
python-dotenv
pydantic
openai
pypdf
pydantic==1.10.13
tiktoken
python-dotenv
bs4
html2text
langchain_openai
langchain-community
uvicorn
pinecone-client
fastapi
langchain-core
langserve
langchain-pinecone
langchain-voyageai
sse_starlette
pymongo
langchain-mongodb
hypercorn
apscheduler
python-jose
python-multipart
bcrypt
passlib
pyjwt
markdown 
passlib
sendgrid
pytube
pypdf2
python-docx
boto3
sqlalchemy
twilio
